+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## v3.0.76 - 2026-09-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v3.0.75...v3.0.76)
+
+- [#7706](https://github.com/ORCID/ORCID-Source/pull/7706): PD-13495 Fixed null mappings for contributors V2 and V3 both works and fundings
+- [#7705](https://github.com/ORCID/ORCID-Source/pull/7705): Error response from /funding endpoint for GET XML request where source is user
+
+## v3.0.75 - 2026-09-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v3.0.74...v3.0.75)
+
+- [#7703](https://github.com/ORCID/ORCID-Source/pull/7703): J21-009 report the invalid enum value instead of a generic bad request
+
 ## v3.0.74 - 2026-09-10
 
 [Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v3.0.73...v3.0.74)